chore: move flux parser into subdir with flux language tooling
