+++ src/timeMachine/components/SubmitQueryButton.test.tsx
@@ -26,7 +26,7 @@ const DELAYTIME = 300
 
 window.TextDecoder = FakeTextDecoder
 
-jest.mock('src/external/parser', () => {
+jest.mock('src/external/languages/flux/parser', () => {
   return {
     parse: jest.fn(getMockedParse()),
   }
