+++ src/variables/utils/hydrateVars.test.ts
@@ -7,7 +7,7 @@ import {
 } from 'src/variables/mocks'
 import {getMockedParse} from 'src/shared/utils/mocks/mockedParse'
 
-jest.mock('src/external/parser', () => {
+jest.mock('src/external/languages/flux/parser', () => {
   return {
     parse: jest.fn(getMockedParse(hydrateVarsAdditionalMappings)),
   }
